Setando endereço do cliente a ser cadastrado.
I created the function that is called when the "Buscar Endereço" button is clicked.
This function will look for this address in the API and use the ID of this address to send the request.

I fix an error in the useFetch hook wich meant that even if there was an error, the setError recieved null.