Creating the logic to save a customer.
The steps to save a customer are:

Check if all inputs are validated. If they're not, the website displays a popup saying that the data is not valid.
If the inputs are validated, the fetch is done and, once completed, feedback will be displayed indicating if the customer has been saved or not.