Creating the logic to update a phone
I created the logic to update a phone
On the phone search screen, when clicking on the "Atualizar Dados" button, the phone screen will be rendered, but with the phone number already filled in and the owner's ID set. The user can change the phone number or the owner.
The system can identify whether the screen is being rendered for registration or updating a phone through the "tableRowSelectedObject" attribute. When its null, is a registry operation, when its filled, is an update operation